Add tests of bridge turn on/off in Network module, minor fixtures refactoring
