+++ openvair/modules/network/tests/api/test_manage.py
@@ -42,6 +42,62 @@ def test_turn_on_interface_success(
     )
 
 
+# TODO: Test this when fixed https://github.com/Aerodisk/openvair/issues/117
+#       (or other tests will fail due to network error)
+# def test_turn_on_bridge_success(
+#         client: TestClient, bridge: Dict
+# ) -> None:
+#     """Test successful bridge interface turn on.
+#
+#     Asserts:
+#     - Response is 200 OK with success message
+#     - Bridge interface power_state becomes 'UNKNOWN' (because it's a bridge)
+#     """
+#     bridge_name = bridge["name"]
+#     client.request('PUT', f'/interfaces/{bridge_name}/turn_off')
+#     wait_for_field_value(
+#         client, f'/interfaces/{bridge["id"]}/', 'power_state', 'DOWN'
+#     )
+#
+#     response = client.request('PUT', f'/interfaces/{bridge_name}/turn_on')
+#     assert response.status_code == status.HTTP_200_OK
+#     assert "turn on command was sent" in response.text.lower()
+#
+#     wait_for_field_value(
+#         client, f'/interfaces/{bridge["id"]}/', 'power_state', 'UNKNOWN'
+#     )
+
+
+def test_turn_on_nonexistent_interface(client: TestClient) -> None:
+    """Test turning on non-existent interface.
+
+    Asserts:
+    - Response is 500 INTERNAL SERVER ERROR
+    - Error message indicates interface not found
+    """
+    nonexistent_name = f'nonexistent-{uuid.uuid4().hex[:8]}'
+
+    response = client.request('PUT', f'/interfaces/{nonexistent_name}/turn_on')
+    assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
+    assert 'not found' in response.text.lower()
+
+
+def test_turn_on_interface_unauthorized(
+    pre_turned_off_interface: Dict, unauthorized_client: TestClient
+) -> None:
+    """Test unauthorized interface turn on.
+
+    Asserts:
+    - Response is 401 UNAUTHORIZED
+    """
+    interface_name = pre_turned_off_interface['name']
+
+    response = unauthorized_client.request(
+        'PUT', f'/interfaces/{interface_name}/turn_on'
+    )
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+
 def test_turn_off_interface_success(
     client: TestClient, pre_turned_on_interface: Dict
 ) -> None:
@@ -65,18 +121,29 @@ def test_turn_off_interface_success(
     )
 
 
-def test_turn_on_nonexistent_interface(client: TestClient) -> None:
-    """Test turning on non-existent interface.
-
-    Asserts:
-    - Response is 500 INTERNAL SERVER ERROR
-    - Error message indicates interface not found
-    """
-    nonexistent_name = f'nonexistent-{uuid.uuid4().hex[:8]}'
-
-    response = client.request('PUT', f'/interfaces/{nonexistent_name}/turn_on')
-    assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
-    assert 'not found' in response.text.lower()
+# TODO: Test this when fixed https://github.com/Aerodisk/openvair/issues/117
+# def test_turn_off_bridge_success(
+#         client: TestClient, bridge: Dict
+# ) -> None:
+#     """Test successful bridge interface turn off.
+#
+#     Asserts:
+#     - Response is 200 OK with success message
+#     - Bridge interface power_state becomes 'DOWN'
+#     """
+#     bridge_name = bridge["name"]
+#     client.request('PUT', f'/interfaces/{bridge_name}/turn_on')
+#     wait_for_field_value(
+#         client, f'/interfaces/{bridge["id"]}/', 'power_state', 'UNKNOWN'
+#     )
+#
+#     response = client.request('PUT', f'/interfaces/{bridge_name}/turn_off')
+#     assert response.status_code == status.HTTP_200_OK
+#     assert "turn off command was sent" in response.text.lower()
+#
+#     wait_for_field_value(
+#         client, f'/interfaces/{bridge["id"]}/', 'power_state', 'DOWN'
+#     )
 
 
 def test_turn_off_nonexistent_interface(client: TestClient) -> None:
@@ -91,58 +158,6 @@ def test_turn_off_nonexistent_interface(client: TestClient) -> None:
     response = client.request('PUT', f'/interfaces/{nonexistent_name}/turn_off')
     assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
     assert 'not found' in response.text.lower()
-
-
-def test_turn_on_interface_twice(
-    client: TestClient, pre_turned_on_interface: Dict
-) -> None:
-    """Test turning on already on interface.
-
-    Asserts:
-    - Response is 200 OK both times
-    - Interface remains 'UP'
-    """
-    interface_name = pre_turned_on_interface['name']
-
-    response = client.request('PUT', f'/interfaces/{interface_name}/turn_on')
-    assert response.status_code == status.HTTP_200_OK
-
-    response = client.get(f'/interfaces/{pre_turned_on_interface["id"]}/')
-    assert response.json()['power_state'] == 'UP'
-
-
-def test_turn_off_interface_twice(
-    client: TestClient, pre_turned_off_interface: Dict
-) -> None:
-    """Test turning off already off interface.
-
-    Asserts:
-    - Response is 200 OK both times
-    - Interface remains 'DOWN'
-    """
-    interface_name = pre_turned_off_interface['name']
-
-    response = client.request('PUT', f'/interfaces/{interface_name}/turn_off')
-    assert response.status_code == status.HTTP_200_OK
-
-    response = client.get(f'/interfaces/{pre_turned_off_interface["id"]}/')
-    assert response.json()['power_state'] == 'DOWN'
-
-
-def test_turn_on_interface_unauthorized(
-    pre_turned_off_interface: Dict, unauthorized_client: TestClient
-) -> None:
-    """Test unauthorized interface turn on.
-
-    Asserts:
-    - Response is 401 UNAUTHORIZED
-    """
-    interface_name = pre_turned_off_interface['name']
-
-    response = unauthorized_client.request(
-        'PUT', f'/interfaces/{interface_name}/turn_on'
-    )
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
 
 def test_turn_off_interface_unauthorized(
